adding single test and restructuring code to be more readable
